파티시스템 1
- 파티 생성 기능
- 파티 찾기
- 파티 HUD
----------------------------
- 파티 탈퇴나 파티에 가입된 플레이어가 나가는 것은 아직 오류남(아직 안 만듦)
- 파티에 뒤늦게 추가된 플레이어가 방을 만든 플레이어에서 동기화가 아직 안댐

+++ Assets/Scripts/Player/PlayerStatDisplay.cs
@@ -27,13 +27,12 @@ public class PlayerStatDisplay : MonoBehaviour
                 {
                     playetCtrl = GameObject.FindGameObjectWithTag("Player").GetComponent<PlayerCtrl>();
                     status = GameObject.FindGameObjectWithTag("Playe;").GetComponent<Status>();
-                }
-                    
+                }  
             }
         }
         else
         {
-            UpdateText(status.HP, playetCtrl.weaponName, status.attackDamage, status.attackSpeed);
+            UpdateText((int)status.HP, playetCtrl.weaponName, status.attackDamage, status.attackSpeed);
         }
     }
